+++ package.json
@@ -6,13 +6,17 @@
   "scripts": {
     "build": "tsc",
     "format": "prettier --write .",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "jest src/"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "devDependencies": {
+    "@jest/globals": "^29.6.2",
+    "jest": "^29.6.2",
+    "jest-environment-jsdom": "^29.6.2",
     "prettier": "^3.0.1",
+    "ts-jest": "^29.1.1",
     "typescript": "^5.1.6"
   }
 }
