+++ src/types/index.d.ts
@@ -1,11 +1,5 @@
 export {};
 
-interface Options {
-  onFormReady: (form: HTMLFormElement) => void;
-  onFormSubmit: (form: HTMLFormElement) => void;
-  onFormSubmitted: (form: HTMLFormElement) => void;
-}
-
 declare global {
   interface Window {
     hbspt: {
